Also generate bindings for some enums used in options
