+++ src-generated/options.rs
@@ -1,0 +1,185 @@
+/* automatically generated by rust-bindgen */
+
+pub const BtorOption_BTOR_OPT_MODEL_GEN: BtorOption = 0;
+pub const BtorOption_BTOR_OPT_INCREMENTAL: BtorOption = 1;
+pub const BtorOption_BTOR_OPT_INCREMENTAL_SMT1: BtorOption = 2;
+pub const BtorOption_BTOR_OPT_INPUT_FORMAT: BtorOption = 3;
+pub const BtorOption_BTOR_OPT_OUTPUT_NUMBER_FORMAT: BtorOption = 4;
+pub const BtorOption_BTOR_OPT_OUTPUT_FORMAT: BtorOption = 5;
+pub const BtorOption_BTOR_OPT_ENGINE: BtorOption = 6;
+pub const BtorOption_BTOR_OPT_SAT_ENGINE: BtorOption = 7;
+pub const BtorOption_BTOR_OPT_AUTO_CLEANUP: BtorOption = 8;
+pub const BtorOption_BTOR_OPT_PRETTY_PRINT: BtorOption = 9;
+pub const BtorOption_BTOR_OPT_EXIT_CODES: BtorOption = 10;
+pub const BtorOption_BTOR_OPT_SEED: BtorOption = 11;
+pub const BtorOption_BTOR_OPT_VERBOSITY: BtorOption = 12;
+pub const BtorOption_BTOR_OPT_LOGLEVEL: BtorOption = 13;
+pub const BtorOption_BTOR_OPT_REWRITE_LEVEL: BtorOption = 14;
+pub const BtorOption_BTOR_OPT_SKELETON_PREPROC: BtorOption = 15;
+pub const BtorOption_BTOR_OPT_ACKERMANN: BtorOption = 16;
+pub const BtorOption_BTOR_OPT_BETA_REDUCE: BtorOption = 17;
+pub const BtorOption_BTOR_OPT_ELIMINATE_SLICES: BtorOption = 18;
+pub const BtorOption_BTOR_OPT_VAR_SUBST: BtorOption = 19;
+pub const BtorOption_BTOR_OPT_UCOPT: BtorOption = 20;
+pub const BtorOption_BTOR_OPT_MERGE_LAMBDAS: BtorOption = 21;
+pub const BtorOption_BTOR_OPT_EXTRACT_LAMBDAS: BtorOption = 22;
+pub const BtorOption_BTOR_OPT_NORMALIZE: BtorOption = 23;
+pub const BtorOption_BTOR_OPT_NORMALIZE_ADD: BtorOption = 24;
+pub const BtorOption_BTOR_OPT_FUN_PREPROP: BtorOption = 25;
+pub const BtorOption_BTOR_OPT_FUN_PRESLS: BtorOption = 26;
+pub const BtorOption_BTOR_OPT_FUN_DUAL_PROP: BtorOption = 27;
+pub const BtorOption_BTOR_OPT_FUN_DUAL_PROP_QSORT: BtorOption = 28;
+pub const BtorOption_BTOR_OPT_FUN_JUST: BtorOption = 29;
+pub const BtorOption_BTOR_OPT_FUN_JUST_HEURISTIC: BtorOption = 30;
+pub const BtorOption_BTOR_OPT_FUN_LAZY_SYNTHESIZE: BtorOption = 31;
+pub const BtorOption_BTOR_OPT_FUN_EAGER_LEMMAS: BtorOption = 32;
+pub const BtorOption_BTOR_OPT_FUN_STORE_LAMBDAS: BtorOption = 33;
+pub const BtorOption_BTOR_OPT_SLS_NFLIPS: BtorOption = 34;
+pub const BtorOption_BTOR_OPT_SLS_STRATEGY: BtorOption = 35;
+pub const BtorOption_BTOR_OPT_SLS_JUST: BtorOption = 36;
+pub const BtorOption_BTOR_OPT_SLS_MOVE_GW: BtorOption = 37;
+pub const BtorOption_BTOR_OPT_SLS_MOVE_RANGE: BtorOption = 38;
+pub const BtorOption_BTOR_OPT_SLS_MOVE_SEGMENT: BtorOption = 39;
+pub const BtorOption_BTOR_OPT_SLS_MOVE_RAND_WALK: BtorOption = 40;
+pub const BtorOption_BTOR_OPT_SLS_PROB_MOVE_RAND_WALK: BtorOption = 41;
+pub const BtorOption_BTOR_OPT_SLS_MOVE_RAND_ALL: BtorOption = 42;
+pub const BtorOption_BTOR_OPT_SLS_MOVE_RAND_RANGE: BtorOption = 43;
+pub const BtorOption_BTOR_OPT_SLS_MOVE_PROP: BtorOption = 44;
+pub const BtorOption_BTOR_OPT_SLS_MOVE_PROP_N_PROP: BtorOption = 45;
+pub const BtorOption_BTOR_OPT_SLS_MOVE_PROP_N_SLS: BtorOption = 46;
+pub const BtorOption_BTOR_OPT_SLS_MOVE_PROP_FORCE_RW: BtorOption = 47;
+pub const BtorOption_BTOR_OPT_SLS_MOVE_INC_MOVE_TEST: BtorOption = 48;
+pub const BtorOption_BTOR_OPT_SLS_USE_RESTARTS: BtorOption = 49;
+pub const BtorOption_BTOR_OPT_SLS_USE_BANDIT: BtorOption = 50;
+pub const BtorOption_BTOR_OPT_PROP_NPROPS: BtorOption = 51;
+pub const BtorOption_BTOR_OPT_PROP_USE_RESTARTS: BtorOption = 52;
+pub const BtorOption_BTOR_OPT_PROP_USE_BANDIT: BtorOption = 53;
+pub const BtorOption_BTOR_OPT_PROP_PATH_SEL: BtorOption = 54;
+pub const BtorOption_BTOR_OPT_PROP_PROB_USE_INV_VALUE: BtorOption = 55;
+pub const BtorOption_BTOR_OPT_PROP_PROB_FLIP_COND: BtorOption = 56;
+pub const BtorOption_BTOR_OPT_PROP_PROB_FLIP_COND_CONST: BtorOption = 57;
+pub const BtorOption_BTOR_OPT_PROP_FLIP_COND_CONST_DELTA: BtorOption = 58;
+pub const BtorOption_BTOR_OPT_PROP_FLIP_COND_CONST_NPATHSEL: BtorOption = 59;
+pub const BtorOption_BTOR_OPT_PROP_PROB_SLICE_KEEP_DC: BtorOption = 60;
+pub const BtorOption_BTOR_OPT_PROP_PROB_CONC_FLIP: BtorOption = 61;
+pub const BtorOption_BTOR_OPT_PROP_PROB_SLICE_FLIP: BtorOption = 62;
+pub const BtorOption_BTOR_OPT_PROP_PROB_EQ_FLIP: BtorOption = 63;
+pub const BtorOption_BTOR_OPT_PROP_PROB_AND_FLIP: BtorOption = 64;
+pub const BtorOption_BTOR_OPT_PROP_NO_MOVE_ON_CONFLICT: BtorOption = 65;
+pub const BtorOption_BTOR_OPT_AIGPROP_USE_RESTARTS: BtorOption = 66;
+pub const BtorOption_BTOR_OPT_AIGPROP_USE_BANDIT: BtorOption = 67;
+pub const BtorOption_BTOR_OPT_QUANT_SYNTH: BtorOption = 68;
+pub const BtorOption_BTOR_OPT_QUANT_DUAL_SOLVER: BtorOption = 69;
+pub const BtorOption_BTOR_OPT_QUANT_SYNTH_LIMIT: BtorOption = 70;
+pub const BtorOption_BTOR_OPT_QUANT_SYNTH_QI: BtorOption = 71;
+pub const BtorOption_BTOR_OPT_QUANT_DER: BtorOption = 72;
+pub const BtorOption_BTOR_OPT_QUANT_CER: BtorOption = 73;
+pub const BtorOption_BTOR_OPT_QUANT_MINISCOPE: BtorOption = 74;
+pub const BtorOption_BTOR_OPT_SORT_EXP: BtorOption = 75;
+pub const BtorOption_BTOR_OPT_SORT_AIG: BtorOption = 76;
+pub const BtorOption_BTOR_OPT_SORT_AIGVEC: BtorOption = 77;
+pub const BtorOption_BTOR_OPT_AUTO_CLEANUP_INTERNAL: BtorOption = 78;
+pub const BtorOption_BTOR_OPT_SIMPLIFY_CONSTRAINTS: BtorOption = 79;
+pub const BtorOption_BTOR_OPT_CHK_FAILED_ASSUMPTIONS: BtorOption = 80;
+pub const BtorOption_BTOR_OPT_CHK_MODEL: BtorOption = 81;
+pub const BtorOption_BTOR_OPT_CHK_UNCONSTRAINED: BtorOption = 82;
+pub const BtorOption_BTOR_OPT_PARSE_INTERACTIVE: BtorOption = 83;
+pub const BtorOption_BTOR_OPT_SAT_ENGINE_LGL_FORK: BtorOption = 84;
+pub const BtorOption_BTOR_OPT_SAT_ENGINE_CADICAL_FREEZE: BtorOption = 85;
+pub const BtorOption_BTOR_OPT_SAT_ENGINE_N_THREADS: BtorOption = 86;
+pub const BtorOption_BTOR_OPT_SIMP_NORMAMLIZE_ADDERS: BtorOption = 87;
+pub const BtorOption_BTOR_OPT_DECLSORT_BV_WIDTH: BtorOption = 88;
+pub const BtorOption_BTOR_OPT_QUANT_SYNTH_ITE_COMPLETE: BtorOption = 89;
+pub const BtorOption_BTOR_OPT_QUANT_FIXSYNTH: BtorOption = 90;
+pub const BtorOption_BTOR_OPT_RW_ZERO_LOWER_SLICE: BtorOption = 91;
+pub const BtorOption_BTOR_OPT_NONDESTR_SUBST: BtorOption = 92;
+pub const BtorOption_BTOR_OPT_NUM_OPTS: BtorOption = 93;
+pub type BtorOption = u32;
+pub const BtorOptBetaReduceMode_BTOR_BETA_REDUCE_NONE: BtorOptBetaReduceMode = 0;
+pub const BtorOptBetaReduceMode_BTOR_BETA_REDUCE_FUN: BtorOptBetaReduceMode = 1;
+pub const BtorOptBetaReduceMode_BTOR_BETA_REDUCE_ALL: BtorOptBetaReduceMode = 2;
+pub type BtorOptBetaReduceMode = u32;
+pub const BtorOptSatEngine_BTOR_SAT_ENGINE_MIN: BtorOptSatEngine = 0;
+pub const BtorOptSatEngine_BTOR_SAT_ENGINE_LINGELING: BtorOptSatEngine = 1;
+pub const BtorOptSatEngine_BTOR_SAT_ENGINE_PICOSAT: BtorOptSatEngine = 2;
+pub const BtorOptSatEngine_BTOR_SAT_ENGINE_MINISAT: BtorOptSatEngine = 3;
+pub const BtorOptSatEngine_BTOR_SAT_ENGINE_CADICAL: BtorOptSatEngine = 4;
+pub const BtorOptSatEngine_BTOR_SAT_ENGINE_CMS: BtorOptSatEngine = 5;
+pub const BtorOptSatEngine_BTOR_SAT_ENGINE_MAX: BtorOptSatEngine = 6;
+pub type BtorOptSatEngine = u32;
+pub const BtorOptEngine_BTOR_ENGINE_MIN: BtorOptEngine = 0;
+pub const BtorOptEngine_BTOR_ENGINE_FUN: BtorOptEngine = 1;
+pub const BtorOptEngine_BTOR_ENGINE_SLS: BtorOptEngine = 2;
+pub const BtorOptEngine_BTOR_ENGINE_PROP: BtorOptEngine = 3;
+pub const BtorOptEngine_BTOR_ENGINE_AIGPROP: BtorOptEngine = 4;
+pub const BtorOptEngine_BTOR_ENGINE_QUANT: BtorOptEngine = 5;
+pub const BtorOptEngine_BTOR_ENGINE_MAX: BtorOptEngine = 6;
+pub type BtorOptEngine = u32;
+pub const BtorOptInputFormat_BTOR_INPUT_FORMAT_MIN: BtorOptInputFormat = 0;
+pub const BtorOptInputFormat_BTOR_INPUT_FORMAT_NONE: BtorOptInputFormat = 1;
+pub const BtorOptInputFormat_BTOR_INPUT_FORMAT_BTOR: BtorOptInputFormat = 2;
+pub const BtorOptInputFormat_BTOR_INPUT_FORMAT_BTOR2: BtorOptInputFormat = 3;
+pub const BtorOptInputFormat_BTOR_INPUT_FORMAT_SMT1: BtorOptInputFormat = 4;
+pub const BtorOptInputFormat_BTOR_INPUT_FORMAT_SMT2: BtorOptInputFormat = 5;
+pub const BtorOptInputFormat_BTOR_INPUT_FORMAT_MAX: BtorOptInputFormat = 6;
+pub type BtorOptInputFormat = u32;
+pub const BtorOptOutputBase_BTOR_OUTPUT_BASE_MIN: BtorOptOutputBase = 0;
+pub const BtorOptOutputBase_BTOR_OUTPUT_BASE_BIN: BtorOptOutputBase = 1;
+pub const BtorOptOutputBase_BTOR_OUTPUT_BASE_HEX: BtorOptOutputBase = 2;
+pub const BtorOptOutputBase_BTOR_OUTPUT_BASE_DEC: BtorOptOutputBase = 3;
+pub const BtorOptOutputBase_BTOR_OUTPUT_BASE_MAX: BtorOptOutputBase = 4;
+pub type BtorOptOutputBase = u32;
+pub const BtorOptOutputFormat_BTOR_OUTPUT_FORMAT_MIN: BtorOptOutputFormat = 0;
+pub const BtorOptOutputFormat_BTOR_OUTPUT_FORMAT_BTOR: BtorOptOutputFormat = 1;
+pub const BtorOptOutputFormat_BTOR_OUTPUT_FORMAT_BTOR2: BtorOptOutputFormat = 2;
+pub const BtorOptOutputFormat_BTOR_OUTPUT_FORMAT_SMT2: BtorOptOutputFormat = 3;
+pub const BtorOptOutputFormat_BTOR_OUTPUT_FORMAT_AIGER_ASCII: BtorOptOutputFormat = 4;
+pub const BtorOptOutputFormat_BTOR_OUTPUT_FORMAT_AIGER_BINARY: BtorOptOutputFormat = 5;
+pub const BtorOptOutputFormat_BTOR_OUTPUT_FORMAT_MAX: BtorOptOutputFormat = 6;
+pub type BtorOptOutputFormat = u32;
+pub const BtorOptDPQsort_BTOR_DP_QSORT_MIN: BtorOptDPQsort = 0;
+pub const BtorOptDPQsort_BTOR_DP_QSORT_JUST: BtorOptDPQsort = 1;
+pub const BtorOptDPQsort_BTOR_DP_QSORT_ASC: BtorOptDPQsort = 2;
+pub const BtorOptDPQsort_BTOR_DP_QSORT_DESC: BtorOptDPQsort = 3;
+pub const BtorOptDPQsort_BTOR_DP_QSORT_MAX: BtorOptDPQsort = 4;
+pub type BtorOptDPQsort = u32;
+pub const BtorOptJustHeur_BTOR_JUST_HEUR_MIN: BtorOptJustHeur = 0;
+pub const BtorOptJustHeur_BTOR_JUST_HEUR_BRANCH_LEFT: BtorOptJustHeur = 1;
+pub const BtorOptJustHeur_BTOR_JUST_HEUR_BRANCH_MIN_APP: BtorOptJustHeur = 2;
+pub const BtorOptJustHeur_BTOR_JUST_HEUR_BRANCH_MIN_DEP: BtorOptJustHeur = 3;
+pub const BtorOptJustHeur_BTOR_JUST_HEUR_MAX: BtorOptJustHeur = 4;
+pub type BtorOptJustHeur = u32;
+pub const BtorOptSLSStrat_BTOR_SLS_STRAT_MIN: BtorOptSLSStrat = 0;
+pub const BtorOptSLSStrat_BTOR_SLS_STRAT_BEST_MOVE: BtorOptSLSStrat = 1;
+pub const BtorOptSLSStrat_BTOR_SLS_STRAT_RAND_WALK: BtorOptSLSStrat = 2;
+pub const BtorOptSLSStrat_BTOR_SLS_STRAT_FIRST_BEST_MOVE: BtorOptSLSStrat = 3;
+pub const BtorOptSLSStrat_BTOR_SLS_STRAT_BEST_SAME_MOVE: BtorOptSLSStrat = 4;
+pub const BtorOptSLSStrat_BTOR_SLS_STRAT_ALWAYS_PROP: BtorOptSLSStrat = 5;
+pub const BtorOptSLSStrat_BTOR_SLS_STRAT_MAX: BtorOptSLSStrat = 6;
+pub type BtorOptSLSStrat = u32;
+pub const BtorOptPropPathSel_BTOR_PROP_PATH_SEL_MIN: BtorOptPropPathSel = 0;
+pub const BtorOptPropPathSel_BTOR_PROP_PATH_SEL_CONTROLLING: BtorOptPropPathSel = 1;
+pub const BtorOptPropPathSel_BTOR_PROP_PATH_SEL_ESSENTIAL: BtorOptPropPathSel = 2;
+pub const BtorOptPropPathSel_BTOR_PROP_PATH_SEL_RANDOM: BtorOptPropPathSel = 3;
+pub const BtorOptPropPathSel_BTOR_PROP_PATH_SEL_MAX: BtorOptPropPathSel = 4;
+pub type BtorOptPropPathSel = u32;
+pub const BtorOptQuantSynth_BTOR_QUANT_SYNTH_MIN: BtorOptQuantSynth = 0;
+pub const BtorOptQuantSynth_BTOR_QUANT_SYNTH_NONE: BtorOptQuantSynth = 1;
+pub const BtorOptQuantSynth_BTOR_QUANT_SYNTH_EL: BtorOptQuantSynth = 2;
+pub const BtorOptQuantSynth_BTOR_QUANT_SYNTH_ELMC: BtorOptQuantSynth = 3;
+pub const BtorOptQuantSynth_BTOR_QUANT_SYNTH_EL_ELMC: BtorOptQuantSynth = 4;
+pub const BtorOptQuantSynth_BTOR_QUANT_SYNTH_ELMR: BtorOptQuantSynth = 5;
+pub const BtorOptQuantSynth_BTOR_QUANT_SYNTH_MAX: BtorOptQuantSynth = 6;
+pub type BtorOptQuantSynth = u32;
+pub use self::BtorOptQuantSynth as BtorOptQuantSynt;
+pub const BtorOptFunEagerLemmas_BTOR_FUN_EAGER_LEMMAS_MIN: BtorOptFunEagerLemmas = 0;
+pub const BtorOptFunEagerLemmas_BTOR_FUN_EAGER_LEMMAS_NONE: BtorOptFunEagerLemmas = 1;
+pub const BtorOptFunEagerLemmas_BTOR_FUN_EAGER_LEMMAS_CONF: BtorOptFunEagerLemmas = 2;
+pub const BtorOptFunEagerLemmas_BTOR_FUN_EAGER_LEMMAS_ALL: BtorOptFunEagerLemmas = 3;
+pub const BtorOptFunEagerLemmas_BTOR_FUN_EAGER_LEMMAS_MAX: BtorOptFunEagerLemmas = 4;
+pub type BtorOptFunEagerLemmas = u32;
+pub const BtorOptIncrementalSMT1_BTOR_INCREMENTAL_SMT1_MIN: BtorOptIncrementalSMT1 = 0;
+pub const BtorOptIncrementalSMT1_BTOR_INCREMENTAL_SMT1_BASIC: BtorOptIncrementalSMT1 = 1;
+pub const BtorOptIncrementalSMT1_BTOR_INCREMENTAL_SMT1_CONTINUE: BtorOptIncrementalSMT1 = 2;
+pub const BtorOptIncrementalSMT1_BTOR_INCREMENTAL_SMT1_MAX: BtorOptIncrementalSMT1 = 3;
+pub type BtorOptIncrementalSMT1 = u32;
