+++ src/lib.rs
@@ -8,3 +8,4 @@
 #![allow(non_camel_case_types, non_snake_case, non_upper_case_globals)]
 
 include!("../src-generated/bindings.rs");
+include!("../src-generated/options.rs");
